requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
deprecation==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-formtools==2.4.1
django-storages==1.14.2
djangorestframework==3.14.0
docopt==0.6.2
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
outcome==1.3.0
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
postgrest==0.11.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
realtime==1.0.0
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
websockets==10.4
wsproto==1.2.0
yarg==0.1.9
